pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kiwoom-playground"
dynamic = ["version"]
description = "Kiwoom REST and WebSocket API SDK with an optional Web Playground"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "requests>=2.32.0",
    "websockets>=12.0",
    "pydantic>=2.13.0"
]

[project.urls]
Homepage = "https://github.com/jaywon99/kiwoom_rest"
Repository = "https://github.com/jaywon99/kiwoom_rest"

[project.optional-dependencies]
playground = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0"
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "types-requests",
    "pre-commit>=3.7.0"
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[project.scripts]
kiwoom-playground = "kiwoom_playground.cli:run_playground"

[tool.setuptools]
packages = ["kiwoom_rest", "kiwoom_playground"]

[tool.setuptools.package-data]
kiwoom_rest = ["apis.json"]
kiwoom_playground = ["templates/*.html"]

[tool.setuptools_scm]
write_to = "kiwoom_rest/_version.py"
fallback_version = "0.0.0"
